Store number of cells in raw/processed bulk vectors
Add a method for slicing columns of a data matrix.

Fix ByteArrayType handling of null blobs.

Add static utilities for generating random EEs and make sure that
ExperimentalDesignRandomUtils is packaged in the test JAR.

Store the NUMBER_OF_CELLS column in a separate table since altering
existing table is very expensive and most rows wouldn't really use the
BLOB anyway.

+++ gemma-core/src/test/java/ubic/gemma/persistence/service/expression/bioAssayData/RandomBulkDataUtils.java
@@ -22,7 +22,7 @@ import java.util.Set;
  */
 public class RandomBulkDataUtils {
 
-    public static void setSeed( int seed ) {
+    public static void setSeed( long seed ) {
         RandomDataUtils.setSeed( seed );
     }
